feat: persist targets in db
This will persist targets in the database 'quietly'. When a target is
added/deleted/updated, the database will be updated. If there is an
error during these persist operations, the error will be logged but will
not return any error to the user. In the case of a target update for a
target we have not persisted yet, we will create the target in the db.